Workspace: Inform about current work state
The Workspace class doesn't have much to do. It simply retreives all the files
from the workspace with relative links to the repository root. The repository
root is defined as the location of the ObjectStore.

TODO Later it should be possible to ignore files and directories from this
list, e.g. the ObjectStore itself.

Signed-off-by: Erik Bernoth <erik.bernoth@gmail.com>